backend/requirements.txt
Flask
flask-cors
Pillow
requests
numpy
albumentations
